Task27 modified

+++ Seminar4/Task27/Program.cs
@@ -1,9 +1,9 @@
 ﻿Console.Clear();
 
 Console.Write("Введите число: ");
-int n = Convert.ToInt32(Console.ReadLine());
+double n = Convert.ToInt32(Console.ReadLine());
 
-int sum = 0;
+double sum = 0;
 int count = 0;
 
 while (n > 0 || n < 0)
